Refactoring: Removed unnecessary code in Repository cards
